IRInterpreter: Various SIMD optimization. Move out the reverse-bits implementation
