+++ Common/BitSet.h
@@ -8,7 +8,21 @@
 #include <cstddef>
 #include "CommonTypes.h"
 
-// Helper functions:
+// TODO: ARM has an intrinsic for the RBIT instruction in some compilers, __rbit.
+inline u32 ReverseBits32(u32 v) {
+	// http://graphics.stanford.edu/~seander/bithacks.html#ReverseParallel
+	// swap odd and even bits
+	v = ((v >> 1) & 0x55555555) | ((v & 0x55555555) << 1);
+	// swap consecutive pairs
+	v = ((v >> 2) & 0x33333333) | ((v & 0x33333333) << 2);
+	// swap nibbles ...
+	v = ((v >> 4) & 0x0F0F0F0F) | ((v & 0x0F0F0F0F) << 4);
+	// swap bytes
+	v = ((v >> 8) & 0x00FF00FF) | ((v & 0x00FF00FF) << 8);
+	// swap 2-byte long pairs
+	v = (v >> 16) | (v << 16);
+	return v;
+}
 
 #ifdef _WIN32
 #include <intrin.h>
@@ -35,6 +49,7 @@ inline int LeastSignificantSetBit(u64 val)
 	_BitScanForward64(&index, val);
 	return (int)index;
 }
+
 #endif
 #else
 inline int CountSetBits(u32 val) { return __builtin_popcount(val); }
